Used pylint to improve PEP8 compliance
Tried to avoid refactoring as much as possible - that can come later.
However, I have disable pylint's C0103 (invalid-name) warning as it
considers anything which isn't inside a function or a class to be a
constant (which these aren't - there are just no functions).

Might need to take another look at how easy the SQL statements are to
read. Is there a nice way to construct very long SQL statements in
Python?